Update 'shutdown' plugin
Lock screen and start screen saver have not worked in ages. Fixes are backwards compatible. Also replaced as much Applescript as I could for now. Plugin devs should try to avoid it as much as possible as Apple has been moving away from it and the requests to control other applications can become tiring. Not to mention the security implications.